Langchain/requirements.txt
langchain
langchain-community
langchain-openai
openai
chromadb
sentence-transformers 